Change default account credentials and remove password setting task
The default account credentials for the Debian build have been updated. The task responsible for setting a password for the default local user in the cleaning phase has been removed as it was found to be unnecessary and potentially insecure. Instead, the password will now be set during the Debian install phase using a provided pre-encrypted password.